Tcl Tool that in turn uses "Beyond Compare" to find missing file
Implemented to try and determine what files are not in a TO
directory of saved records but are in a FROM directory.

I can then manually copy/move the 'missing' file(s) to the
TO directory.

Tried to make it very generic but may be rather tied to my
own 'records' areas.

This was sort of working a year ago ... but I haven't used it
since.  But I want to capture this 'so far' work.

 Changes to be committed:
	new file:   do_diff.bat
	new file:   find_missing_files.tcl
	new file:   get_full_file_info.bat
	new file:   get_short_file_name.bat