Fix file prolog.
As best I remember, I used 'gnat' on CentOS, then CentOS crashed
on me during an 'upgrade', so I went to Ubuntu.  Then, the
'gcc' compiler (on Ubuntu) said it had an Ada option.  I tried
that but it bombed.

So I gave up.  This was all about June, 2024.

So, still on Ubuntu (and debating if I should setup a SunOS
installation since Oracle says it's 'free'), I did:

    sudo apt install gnat

And that worked, brought down GNU gnat toolset and that brought
me back to what I had been doing on CentOS.

Changes to be committed:
	modified:   hello_world.adb